Issue #28 - AwsKinesisTransactionStoreFactory is responsible for stream creation
