+++ iron-inttest/src/test/java/io/axway/iron/spi/aws/BaseInttest.java
@@ -3,14 +3,11 @@ package io.axway.iron.spi.aws;
 import java.io.*;
 import java.nio.file.Paths;
 import java.util.*;
-import java.util.function.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.annotations.BeforeClass;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.services.kinesis.AmazonKinesis;
-import com.amazonaws.services.kinesis.model.DescribeStreamRequest;
-import com.amazonaws.services.kinesis.model.DescribeStreamResult;
 import com.amazonaws.services.kinesis.model.LimitExceededException;
 import com.amazonaws.services.kinesis.model.ResourceNotFoundException;
 import com.amazonaws.services.s3.AmazonS3;
@@ -23,16 +20,15 @@ import io.axway.iron.spi.aws.s3.AwsS3Utils;
 
 import static com.amazonaws.SDKGlobalConfiguration.*;
 import static io.axway.iron.spi.aws.AwsProperties.*;
+import static io.axway.iron.spi.aws.AwsUtils.performAmazonActionWithRetry;
 import static io.axway.iron.spi.aws.kinesis.AwsKinesisUtils.*;
 
 /**
  * To run these tests, localstack must be started with Kinesis and S3, and localstack profile must be enabled.
  */
 public abstract class BaseInttest {
-    private static final int MIN_3 = 1000 * 60 * 3;
+
     private static final Logger LOG = LoggerFactory.getLogger(BaseInttest.class);
-    private static final int DEFAULT_RETRY_DURATION_IN_MILLIS = 5000;
-    private static final int DEFAULT_RETRY_COUNT = 5;
     private static final String AWS_KINESIS_STREAM_NAME_PREFIX = "condor-iron-transaction-store-";
 
     protected final Properties m_configuration = loadConfiguration();
@@ -92,8 +88,7 @@ public abstract class BaseInttest {
     protected void createStreamAndWaitActivation(String storeName) {
         AmazonKinesis amazonKinesis = buildKinesisClient(m_configuration);
         String streamName = AWS_KINESIS_STREAM_NAME_PREFIX + storeName;
-        createStreamIfNotExists(amazonKinesis, streamName, 1);
-        waitStreamActivation(amazonKinesis, streamName, MIN_3);
+        ensureStreamExists(amazonKinesis, streamName, LOG);
     }
 
     protected void deleteKinesisStream(String storeName) {
@@ -144,25 +139,6 @@ public abstract class BaseInttest {
     }
 
     /**
-     * Create a stream if it does not already exist.
-     *
-     * @param kinesis AmazonKinesis client
-     * @param streamName the name of the stream
-     * * @throws LimitExceededException after retry exhausted
-     */
-    private static void createStreamIfNotExists(AmazonKinesis kinesis, String streamName, int shardCount) {
-        performAmazonActionWithRetry("createStream", () -> {
-            DescribeStreamRequest describeStreamRequest = new DescribeStreamRequest().withStreamName(streamName).withLimit(1);
-            try {
-                kinesis.describeStream(describeStreamRequest);
-            } catch (ResourceNotFoundException e) {
-                kinesis.createStream(streamName, shardCount);
-            }
-            return null;
-        }, DEFAULT_RETRY_COUNT, DEFAULT_RETRY_DURATION_IN_MILLIS);
-    }
-
-    /**
      * Delete stream silently
      * <p>
      * Won't raise an error if stream doesn't exists
@@ -184,62 +160,6 @@ public abstract class BaseInttest {
                 // No need to delete resource doesn't even exists
             }
             return null;
-        }, DEFAULT_RETRY_COUNT, DEFAULT_RETRY_DURATION_IN_MILLIS);
-    }
-
-    /**
-     * Waits that the stream has been created.
-     */
-    private static void waitStreamActivation(AmazonKinesis consumer, String streamName, long streamCreationTimeoutMillis) {
-        DescribeStreamRequest describeStreamRequest = new DescribeStreamRequest().withStreamName(streamName).withLimit(1);
-        DescribeStreamResult describeStreamResult = null;
-        String streamStatus = null;
-        long endTime = System.currentTimeMillis() + streamCreationTimeoutMillis;
-        do {
-            try {
-                describeStreamResult = consumer.describeStream(describeStreamRequest);
-                streamStatus = describeStreamResult.getStreamDescription().getStreamStatus();
-                if (streamStatus.equals(ACTIVE_STREAM_STATUS)) {
-                    break;
-                }
-                try {
-                    Thread.sleep(100);
-                } catch (Exception ignored) {
-                }
-            } catch (ResourceNotFoundException ignored) {
-            }
-        } while (System.currentTimeMillis() < endTime);
-        if (describeStreamResult == null || streamStatus == null || !streamStatus.equals(ACTIVE_STREAM_STATUS)) {
-            throw new AwsKinesisException("Stream never went active",
-                                          args -> args.add("streamName", streamName).add("streamCreationTimeoutMillis", streamCreationTimeoutMillis));
-        }
-    }
-
-    /**
-     * Handle retry for amazon quotas
-     *
-     * @param actionLabel action label used for logging purpose only
-     * @param action the action to retry
-     * @param retryLimit retry number limit
-     * @param durationInMillis duration between each retry
-     * @throws LimitExceededException after retry exhausted
-     */
-    private static void performAmazonActionWithRetry(String actionLabel, Supplier<Void> action, int retryLimit, int durationInMillis) {
-        int retryCount = 0;
-        do {
-            try {
-                action.get();
-                return;
-            } catch (LimitExceededException lee) {
-                // We should just wait a little time before trying again
-                LOG.debug("LimitExceededException while doing " + actionLabel + " will retry " + (retryLimit - retryCount) + " times");
-            }
-            try {
-                Thread.sleep(durationInMillis);
-                LOG.debug("Throttling {} for {} ms", actionLabel, durationInMillis);
-            } catch (InterruptedException ignored) {
-            }
-        } while (retryCount++ < retryLimit);
-        throw new LimitExceededException("Can't do " + actionLabel + " after " + retryLimit + " retries");
+        }, DEFAULT_RETRY_COUNT, DEFAULT_RETRY_DURATION_IN_MILLIS, LOG);
     }
 }
