+++ iron-spi-aws/src/main/java/io/axway/iron/spi/aws/AwsUtils.java
@@ -1,9 +1,12 @@
 package io.axway.iron.spi.aws;
 
 import java.util.*;
+import java.util.function.*;
+import org.slf4j.Logger;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder;
+import com.amazonaws.services.kinesis.model.LimitExceededException;
 
 import static io.axway.iron.spi.aws.AwsProperties.*;
 import static io.axway.iron.spi.aws.PropertiesHelper.getValue;
@@ -28,4 +31,34 @@ public class AwsUtils {
             }
         }
     }
+
+    /**
+     * Handle retry for amazon quotas
+     *
+     * @param actionLabel action label used for logging purpose only
+     * @param action the action to retry
+     * @param retryLimit retry number limit
+     * @param durationInMillis duration between each retry
+     * @param logger logger
+     * @throws LimitExceededException after retry exhausted
+     */
+    public static void performAmazonActionWithRetry(String actionLabel, Supplier<Void> action, int retryLimit, int durationInMillis, Logger logger) {
+        int retryCount = 0;
+        do {
+            try {
+                action.get();
+                return;
+            } catch (LimitExceededException lee) {
+                // We should just wait a little time before trying again
+                logger.debug("LimitExceededException while doing " + actionLabel + " will retry " + (retryLimit - retryCount) + " times");
+            }
+            try {
+                Thread.sleep(durationInMillis);
+                logger.debug("Throttling {} for {} ms", actionLabel, durationInMillis);
+            } catch (InterruptedException ignored) {
+            }
+        } while (retryCount++ < retryLimit);
+        throw new LimitExceededException("Can't do " + actionLabel + " after " + retryLimit + " retries");
+    }
+
 }
